Added caching to TimberLoader::render
Added transient, site-transient and object cache modes to TimberLoader::render and Timber::render methods. Expiration can differ between logged in and non-logged in users (primarily to disable caching for logged in users). Uses data + template name as the cache key so the same template can be loaded with different data and so that a data change automatically invalidates cache data.