update timber-comment.php to have avatars
Ok, take 2. This is working well for me.  Tested it with the various default settings and email addresses with & without gravatars. There are no tests though. not 100% sure how i would do that but willing to have a go. Do you have a test suite for this?

Let me know if i need to tidy it up / its generally no good / any other feedback.