Add Timber::maybe_set_preview and call it in Timber::get_posts
Added the method maybe_set_preview, largely based on WPs internal _set_preview function. Checks if the GET params call for a preview and if the current user is allowed to edit the current post. It scans the posts array for the previewed post and if present replaces it by the last autosave by the current user (again, if available).