Made Timber autoload dependencies using Composer generated autoloader
Added include for `vendor/autoload.php` at the start of main plugin file.
Removed manual include and init of Twig's own autoloader and manual include of Router.
Rewrote unit test for Router class to simply assert that the class is available (instead of checking that it's unavailable before registering the first route)