Only render the twig template on template_include ( = WooCommerce compatibility)
Twig templates are now returned as templates into the WP template loading system, and caught just in time by a filter on template include, that checks if the to-be-included file is a twig template.

With index templates, the output buffer is not flushed directly, to give other plugins a chance to set their template. If they don't it is automatically outputted by TimberTemplateLoader::template_include. If they do, the output buffer is silently cleaned.