Rename init() to create_twig_loader(), and move cache filters to controller.
The Twig Loader is accessible via get_loader(), and exposing the
$loader property makes the class vulnerable to alterations after
initialization. Even via get_loader() it is possible to modify the
loader, but not replacing the loader object, which could be done via
the WordPress hooks already.

Renaming the new init() method to create_twig_loader(), and modifying
it to return a newly created Twig Loader object, which the controller
stores in the private $loader property once and for all, internal
integrity is insured and overloading made possible…