Change filters into actions
The Twig environment object is based to the filters, which can be
manipulated by both actions and filters. However filters can replace
the object, which may break the filter chain altogether. By changing
the filters into actions this is no longer the case. Unless it was the
intention to replace the environment in the first place, actions should
have been used in the first place. Internally in wordpress actions is
actually filters, which does not return anything, so backward
compatibility should be intact.