Add AllowDynamicProperties attribute
Dynamic (non-explicitly declared) properties are deprecated as of PHP 8.2 and are expected to become a fatal error in PHP 9.0.

To reduce noise, both in error logs and test run logs, this commit adds the `#[AllowDynamicProperties]` attribute to Timber's classes.

Ideally, a magic `__set()` method should be added which will require adding a catch-all data property to store the non-declared fields.